moves ping to own package; adds unit and api tests
